# solutions/partie-11/test/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/inscription": {
      "post": {
        "operationId": "AuthController_inscrire",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/connexion": {
      "post": {
        "operationId": "AuthController_connecter",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnexionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/moi": {
      "get": {
        "operationId": "AuthController_moi",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/rafraichir": {
      "post": {
        "operationId": "AuthController_rafraichir",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RafraichirDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/deconnexion": {
      "post": {
        "operationId": "AuthController_deconnecter",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/vendeurs": {
      "get": {
        "operationId": "VendeursController_trouverTous",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      },
      "post": {
        "operationId": "VendeursController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerVendeurDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      }
    },
    "/api/vendeurs/{id}": {
      "get": {
        "operationId": "VendeursController_trouverUn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      },
      "patch": {
        "operationId": "VendeursController_renommer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerVendeurDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      },
      "delete": {
        "operationId": "VendeursController_supprimer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      }
    },
    "/api/vendeurs/{id}/produits": {
      "post": {
        "operationId": "VendeursController_ajouterProduit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerProduitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "vendeurs"
        ]
      }
    },
    "/api/commandes": {
      "post": {
        "operationId": "CommandesController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerCommandeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commandes"
        ]
      }
    },
    "/api/commandes/{id}": {
      "get": {
        "operationId": "CommandesController_trouverUne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commandes"
        ]
      }
    },
    "/api/commandes/{id}/expedier": {
      "post": {
        "operationId": "CommandesController_expedier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "commandes"
        ]
      }
    },
...
    "/api/vendeurs/{vendeurId}/imports": {
      "post": {
        "operationId": "ImportsController_importer",
        "parameters": [
          {
            "name": "vendeurId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImporterCatalogueDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "imports"
        ]
      }
    },
    "/api/imports/{importId}": {
      "get": {
        "operationId": "ImportsController_etat",
        "parameters": [
          {
            "name": "importId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "imports"
        ]
      }
    }
  },
  "info": {
    "title": "API Marketplace",
    "description": "Catalogue de produits, vendeurs, commandes et comptes de la marketplace.",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "InscriptionDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse e-mail (unique)",
            "example": "alice@exemple.fr"
          },
          "motDePasse": {
            "type": "string",
            "description": "Mot de passe, 8 caractères au moins",
            "minLength": 8,
            "example": "MotDePasse!42"
          }
        },
        "required": [
          "email",
          "motDePasse"
        ]
      },
      "ConnexionDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse e-mail du compte",
            "example": "alice@exemple.fr"
          },
          "motDePasse": {
            "type": "string",
            "description": "Mot de passe du compte",
            "example": "MotDePasse!42"
          }
        },
        "required": [
          "email",
          "motDePasse"
        ]
      },
      "RafraichirDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Le refresh token reçu à la connexion"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreerProduitDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "Nom du produit",
            "example": "Lampe de bureau"
          },
          "prix": {
            "type": "number",
            "description": "Prix unitaire en euros",
            "minimum": 0,
            "example": 30
          },
          "categorie": {
            "type": "string",
            "enum": [
              "papeterie",
              "informatique",
              "mobilier"
            ],
            "example": "mobilier"
          },
          "prixAchat": {
            "type": "number",
            "description": "Prix d'achat en euros (visible du propriétaire et d'un admin)",
            "minimum": 0,
            "example": 18
          }
        },
        "required": [
          "nom",
          "prix",
          "categorie"
        ]
      },
...
      "CreerVendeurDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "Nom de la boutique",
            "example": "Atelier du Bois"
          }
        },
        "required": [
          "nom"
        ]
      },
      "CreerCommandeDto": {
        "type": "object",
        "properties": {
          "lignes": {
            "description": "Les lignes de la commande",
            "example": [
              {
                "varianteId": 1,
                "quantite": 2
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "lignes"
        ]
      },
...
      "ImporterCatalogueDto": {
        "type": "object",
        "properties": {
          "csv": {
            "type": "string",
            "description": "Le catalogue au format CSV (séparateur `;`), en-tête compris",
            "example": "nom;prix;categorie\nTable;120;mobilier"
          }
        },
        "required": [
          "csv"
        ]
      }
    }
  }
}